backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
python-dotenv
asyncpg